UNI-328 #comment Switch everywhere from JS Date() to Moment() instantiation. Rename "Unicorn" => "HTMStudio" in docs and other simple places. Alphabetize lists, etc.
