+++ unicorn/js/browser/components/Main.jsx
@@ -28,9 +28,9 @@ import ThemeManager from 'material-ui/lib/styles/theme-manager';
 import FileValidateAction from '../actions/FileValidate';
 import FileList from '../components/FileList';
 import FileDetails from '../components/FileDetails';
+import HTMStudioTheme from '../lib/MaterialUI/HTMStudioTheme';
 import LeftNav from '../components/LeftNav';
 import ModelList from '../components/ModelList';
-import UnicornTheme from '../lib/MaterialUI/UnicornTheme';
 
 const app = remote.app;
 const dialog = remote.require('dialog');
@@ -47,7 +47,7 @@ const dialog = remote.require('dialog');
   getModelClient: React.PropTypes.func,
   getParamFinderClient: React.PropTypes.func
 })
-@ThemeDecorator(ThemeManager.getMuiTheme(UnicornTheme)) // eslint-disable-line new-cap
+@ThemeDecorator(ThemeManager.getMuiTheme(HTMStudioTheme)) // eslint-disable-line new-cap
 export default class Main extends React.Component {
 
   static get contextTypes() {
