+++ unicorn/js/main/TimeAggregator.js
@@ -15,9 +15,9 @@
 //
 // http://numenta.org/licenses/
 
-
-var util = require('util'); // eslint-disable-line no-var
+var moment = require('moment');
 var Transform = require('stream').Transform; // eslint-disable-line no-var
+var util = require('util'); // eslint-disable-line no-var
 
 
 /**
@@ -45,7 +45,7 @@ util.inherits(TimeAggregator, Transform);
 
 TimeAggregator.prototype._transform = function (data, encoding, done) {
   if (this._timefield in data) {
-    let timestamp = new Date(data[this._timefield]).getTime();
+    let timestamp = moment(data[this._timefield]).valueOf();
     if (this._timebucket === 0) {
       this._timebucket = timestamp + this._interval;
     } else if (timestamp >= this._timebucket) {
