+++ unicorn/scripts/OSX/package.json
@@ -1,7 +1,7 @@
 {
   "name": "portable_python_darwin",
   "version": "1.0.0",
-  "description": "Portable Python for Unicorn (Mac OSX)",
+  "description": "Portable Python for HTM Studio (Mac OSX)",
   "main": "index.js",
   "os": [
     "darwin"
